requirements.txt
google-cloud-aiplatform
google-cloud-bigquery
google-adk
Jinja2
python-dotenv
Flask